requirements.txt
#------- Core dependencies -------
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0

transformers[sentencepiece]>=4.33,<4.34
rich>=13.0.0,<14.0.0
scikit-learn>=1.3,<1.4
overrides>=7.4,<7.5
art==6.1
pprintpp==0.4.0
colorama==0.4.6

#------- Optional dependencies -------

# train
lightning>=2.0,<2.1
datasets>=2.13,<2.15 
hydra-core>=1.3,<1.4
hydra_colorlog
wandb>=0.15,<0.16

# faiss
faiss-cpu==1.7.4 # needed by: faiss

# serve
fastapi>=0.104,<0.105 # needed by: serve, ray
uvicorn[standard]==0.23.2 # needed by: serve, ray
gunicorn==21.2.0 # needed by: serve, ray
spacy>=3.7,<3.8 # needed by: serve, ray
streamlit>=1.28,<1.29 # needed by: serve, ray
streamlit_extras>=0.3,<0.4 # needed by: serve, ray
pyvis # needed by: serve, ray
ray[serve]>=2.7,<=2.8 # needed by: ray

# retriever

# reader

# dev
pre-commit # needed by: dev
black[d] # needed by: dev
isort # needed by: dev
